add concurrent test

+++ conf/evolutions/default/1.sql
@@ -6,7 +6,7 @@
 create table group_member (
   id                            varchar(255) not null,
   uid                           varchar(255),
-  gid                           integer,
+  gid                           bigint,
   constraint pk_group_member primary key (id)
 );
